Adding Chronobreak (#377)
https://github.com/rollersteaam/chronobreak

I think my app Chronobreak should be added because it's a useful Pomodoro timer that works surprisingly differently compared to other timers of the same ink, which is the exact reason I bothered making it in the first place. Many things like audible stimuli from setting the timer into place, and the more enforced segments for breaks and work times as phases makes getting motivated to start work on a daunting task significantly easier.