feat: Update repository url for my exist app : Serina (#1435)
Hi, I lost access to my GitHub account when I moved to Canada because of double authentication. So I created a copy of my original repository to resume the work I've been leaving on hold all this time. With these changes it allows people to access future new versions of the application that will be released soon. Thanks in advance.

Co-authored-by: Vlad Hashimoto <hashimoto.stream@gmail.com>